backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
pydantic
numpy
pillow
openaireportlab==4.4.5
